editor/libeditor/tests/test_cmd_paragraphState.html

<!doctype html>
<html>
<head>
<meta charset="utf-8">
<title>Testing "cmd_paragraphState" behavior</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<div contenteditable></div>
<script>
"use strict";

SimpleTest.waitForExplicitFinish();
SimpleTest.waitForFocus(() => {
  const editor = document.querySelector("div[contenteditable]");

  editor.innerHTML = "<div><p>abc</p></div>";
  editor.focus();
  getSelection().collapse(editor.querySelector("p").firstChild, 1);
  editor.getBoundingClientRect();
  SpecialPowers.doCommand(window, "cmd_paragraphState", "");
  is(
    editor.innerHTML,
    "<div>abc</div>",
    "cmd_paragraphState with empty string should remove the parent block element"
  );

  editor.innerHTML = "<div><div contenteditable=\"false\"><p contenteditable>abc</p></div></div>";
  editor.focus();
  getSelection().collapse(editor.querySelector("p").firstChild, 1);
  editor.getBoundingClientRect();
  SpecialPowers.doCommand(window, "cmd_paragraphState", "");
  is(
    editor.innerHTML,
    "<div><div contenteditable=\"false\"><p contenteditable=\"\">abc</p></div></div>",
    "cmd_paragraphState with empty string should not remove editing host"
  );

  editor.innerHTML = "<div><div contenteditable=\"false\"><p><span contenteditable>abc</span></p></div></div>";
  editor.focus();
  getSelection().collapse(editor.querySelector("span").firstChild, 1);
  editor.getBoundingClientRect();
  SpecialPowers.doCommand(window, "cmd_paragraphState", "");
  is(
    editor.innerHTML,
    "<div><div contenteditable=\"false\"><p><span contenteditable=\"\">abc</span></p></div></div>",
    "cmd_paragraphState with empty string should not remove parents of inline editing host"
  );

  SimpleTest.finish();
});
</script>
</body>
</html>
